Migrate from MyBatis to Spring JPA for Session
- Use type Long for Session#id

+++ src/main/java/de/sustineo/simdesk/entities/LeaderboardLine.java
@@ -14,7 +14,7 @@ import java.util.List;
 @NoArgsConstructor
 public class LeaderboardLine {
     private Integer id;
-    private Integer sessionId;
+    private Long sessionId;
     private Integer ranking;
     private AccCupCategory cupCategory;
     private Integer carId;
